+++ kamehouse-groot/public/kame-house-groot/admin/my-scripts/exec-script.php
@@ -23,7 +23,7 @@
 </head>
 
 <body>
-  <div id="groot-menu-wrapper"></div>
+  <div id="groot-menu-wrapper" onmouseover="header.showGrootMenu()" onmouseleave="header.hideGrootMenu()"></div>
   <div class="banner-wrapper">
     <div id="banner" class="fade-in-out-15s banner-fuji">
       <div class="default-layout banner-text">